requirements.txt
Django==2.2.1
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
django-rest-swagger==2.2.0
PyJWT==1.7.1
pytz==2019.1
psycopg2==2.7.5
celery==4.2.1